refactor(user-navbar): extract and customize user navbar component
- Move Navbar component from shared UI library to user app components
- Rename Navbar to UserNavbar and remove appType prop
- Simplify navigation links to user-specific routes only
- Update user app layout to use UserNavbar instead of shared Navbar
- Adjust import paths for type declarations in admin and user apps
- Update UserMenu to redirect to homepage on logout instead of login page
- Remove Navbar export from shared UI library to decouple user-specific UI
