+++ apps/user/src/components/navbar/index.ts
@@ -1,0 +1,1 @@
+export { UserNavbar } from './navbar';
